Extract: TYPE_MAP to own file
