+++ lib/create-matcher.js
@@ -7,7 +7,6 @@ var functionName = require("@sinonjs/commons").functionName;
 var get = require("lodash").get;
 var iterableToString = require("./iterable-to-string");
 var objectProto = require("@sinonjs/commons").prototypes.object;
-var stringProto = require("@sinonjs/commons").prototypes.string;
 var typeOf = require("@sinonjs/commons").typeOf;
 var valueToString = require("@sinonjs/commons").valueToString;
 
@@ -16,67 +15,16 @@ var assertMethodExists = require("./create-matcher/assert-method-exists");
 var assertType = require("./create-matcher/assert-type");
 var isIterable = require("./create-matcher/is-iterable");
 var isMatcher = require("./create-matcher/is-matcher");
-var matchObject = require("./create-matcher/match-object");
+
 var matcherPrototype = require("./create-matcher/matcher-prototype");
 
 var arrayIndexOf = arrayProto.indexOf;
-var join = arrayProto.join;
-var map = arrayProto.map;
 var some = arrayProto.some;
 
 var hasOwnProperty = objectProto.hasOwnProperty;
 var objectToString = objectProto.toString;
 
-var stringIndexOf = stringProto.indexOf;
-
-var TYPE_MAP = {
-    function: function(m, expectation, message) {
-        m.test = expectation;
-        m.message = message || "match(" + functionName(expectation) + ")";
-    },
-    number: function(m, expectation) {
-        m.test = function(actual) {
-            // we need type coercion here
-            return expectation == actual; // eslint-disable-line eqeqeq
-        };
-    },
-    object: function(m, expectation) {
-        var array = [];
-
-        if (typeof expectation.test === "function") {
-            m.test = function(actual) {
-                return expectation.test(actual) === true;
-            };
-            m.message = "match(" + functionName(expectation.test) + ")";
-            return m;
-        }
-
-        array = map(Object.keys(expectation), function(key) {
-            return key + ": " + valueToString(expectation[key]);
-        });
-
-        m.test = function(actual) {
-            return matchObject(actual, expectation);
-        };
-        m.message = "match(" + join(array, ", ") + ")";
-
-        return m;
-    },
-    regexp: function(m, expectation) {
-        m.test = function(actual) {
-            return typeof actual === "string" && expectation.test(actual);
-        };
-    },
-    string: function(m, expectation) {
-        m.test = function(actual) {
-            return (
-                typeof actual === "string" &&
-                stringIndexOf(actual, expectation) !== -1
-            );
-        };
-        m.message = 'match("' + expectation + '")';
-    }
-};
+var TYPE_MAP = require("./create-matcher/type-map");
 
 /**
  * Creates a matcher object for the passed expectation
